+++ instrumentation/go.mongodb.org/mongo-driver/mongo/otelmongo/doc.go
@@ -6,10 +6,18 @@
 // This package is compatible with v0.2.0 of
 // go.mongodb.org/mongo-driver/mongo.
 //
-// `NewMonitor` will return an event.CommandMonitor which is used to trace
+// NewMonitor will return an event.CommandMonitor which is used to trace
 // requests.
 //
 // This code was originally based on the following:
-// - https://github.com/DataDog/dd-trace-go/tree/02f0449efa3cb382d499fadc873957385dcb2192/contrib/go.mongodb.org/mongo-driver/mongo
-// - https://github.com/DataDog/dd-trace-go/tree/v1.23.3/ddtrace/ext
+//   - https://github.com/DataDog/dd-trace-go/tree/02f0449efa3cb382d499fadc873957385dcb2192/contrib/go.mongodb.org/mongo-driver/mongo
+//   - https://github.com/DataDog/dd-trace-go/tree/v1.23.3/ddtrace/ext
+//
+// The "OTEL_SEMCONV_STABILITY_OPT_IN" environment variable can be used to opt
+// into semconv/v1.26.0:
+//   - "database": emit v1.26.0 semantic conventions
+//   - "": emit v1.21.0 (default) semantic conventions
+//   - "database/dup": emit v1.21.0 (default) and v1.26.0
+//
+// By default, otelmongo only emits v1.21.0.
 package otelmongo // import "go.opentelemetry.io/contrib/instrumentation/go.mongodb.org/mongo-driver/mongo/otelmongo"